+++ src/main/java/org/bot/nullbot/component/ai/DeepSeekClient.java
@@ -37,6 +37,7 @@ import org.bot.nullbot.entity.CommandEvent;
 import org.bot.nullbot.entity.EmbeddedCommandEvent;
 import org.bot.nullbot.service.SettingService;
 import org.bot.nullbot.util.Base64Util;
+import org.bot.nullbot.util.BotCtxUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Lazy;
@@ -140,7 +141,7 @@ public class DeepSeekClient {
             String message, Bot bot, Event event
     ) throws Exception {
 
-        SettingPO setting = settingService.get(groupId);
+        SettingPO setting = BotCtxUtil.getSetting();
 
         if (setting.isAntiInjection()) {
             String req = """
@@ -223,7 +224,7 @@ public class DeepSeekClient {
      * @return 清除模式
      */
     public ChatScope clearGroupHistory(Long groupId, Long userId) {
-        ChatScope chatScope = settingService.get(groupId).getChatScope();
+        ChatScope chatScope = BotCtxUtil.getSetting().getChatScope();
         switch (chatScope) {
             case Group -> chatStorage.clearGroupHistory(groupId);
             case Personal -> {
@@ -242,8 +243,7 @@ public class DeepSeekClient {
      * @return 聊天记录列表
      */
     public List<ChatMessage> getGroupHistory(Long groupId, Long userId) {
-        SettingPO setting = settingService.get(groupId);
-        return switch (setting.getChatScope()) {
+        return switch (BotCtxUtil.getSetting().getChatScope()) {
             case Group -> chatStorage.getGroupHistory(groupId);
             case Personal -> chatStorage.getUserHistory(userId);
             case Monitor -> chatStorage.getMonitorHistory(groupId);